+++ src/util/util.ts
@@ -66,6 +66,6 @@ export const minmax = (data: Array<number>) => data.reduce((acc: Array<number>, 
 
 // helper function: async wait
 export async function wait(time) {
-  const waiting = new Promise((resolve) => setTimeout(() => resolve(true), time));
+  const waiting = new Promise((resolve) => { setTimeout(() => resolve(true), time); });
   await waiting;
 }